+++ runner/limit_windows.go
@@ -1,0 +1,6 @@
+// +build windows
+
+package runner
+
+func initLimit() {
+}
