+++ runner/settings.go
@@ -5,6 +5,7 @@ import (
 	"github.com/pilu/config"
 	"os"
 	"path/filepath"
+	"runtime"
 	"strconv"
 	"strings"
 	"time"
@@ -116,7 +117,11 @@ func buildName() string {
 	return settings["build_name"]
 }
 func buildPath() string {
-	return filepath.Join(tmpPath(), buildName())
+	p := filepath.Join(tmpPath(), buildName())
+	if runtime.GOOS == "windows" && filepath.Ext(p) != ".exe" {
+		p += ".exe"
+	}
+	return p
 }
 
 func buildErrorsFileName() string {
